fix: 💫 add support for exact versions
