+++ Cargo.toml
@@ -22,10 +22,10 @@ ahash = "0.8.11"
 anyhow = "1.0.95"
 clap = { version = "4.5.23", features = ["derive"] }
 clap-verbosity-flag = "3.0.2"
-config = "0.15.4"
+config = "0.15.6"
 env_logger = "0.11.6"
 git2 = "0.20.0"
-log = "0.4.22"
+log = "0.4.25"
 nom = "7.1.3"
 semver = "1.0.24"
 serde = { version = "1.0.217", features = ["derive"] }
